requirements.txt
streamlit
torch==2.5.1
torchvision==0.20.1
opencv-python-headless
Pillow
numpy